토í큰 ì쿠키로 저장
