+++ apps/package.json
@@ -4,6 +4,7 @@
   "private": true,
   "scripts": {
     "dev": "next dev -p 3000",
+    "dev:https": "next dev -p 3000 --experimental-https",
     "build": "next build",
     "start": "next start",
     "lint": "next lint"
