--- apps/src/app/profile/page.tsx
@@ -1,5 +1,0 @@
-import UserProfileTemp from "@/components/UserProfile-temp";
-
-export default function ProfilePage() {
-    return <UserProfileTemp />;
-}
