+++ apps/src/app/providers.tsx
@@ -1,8 +1,7 @@
 'use client';
 
 import { SessionProvider } from 'next-auth/react';
-import { AuthProvider } from '@/contexts/AuthContext';
 
 export function Providers({ children }: { children: React.ReactNode }) {
-  return <SessionProvider> <AuthProvider>{children} </AuthProvider></SessionProvider>;
+  return <SessionProvider>{children}</SessionProvider>;
 }
