--- apps/src/contexts/AuthContext.tsx
@@ -1,30 +1,0 @@
-'use client';
-import { createContext, useContext, ReactNode } from 'react';
-import { useAuth } from '@/hooks/useAuth';
-
-interface AuthContextType {
-    token: string | null;
-    isAuthenticated: boolean;
-    login: (token: string) => void;
-    logout: () => void;
-}
-
-const AuthContext = createContext<AuthContextType | undefined>(undefined);
-
-export const useAuthContext = () => {
-    const context = useContext(AuthContext);
-    if (context === undefined) {
-        throw new Error('useAuthContext must be used within an AuthProvider');
-    }
-    return context;
-};
-
-export function AuthProvider({ children }: { children: ReactNode }) {
-    const auth = useAuth();
-
-    return (
-        <AuthContext.Provider value={auth}>
-            {children}
-        </AuthContext.Provider>
-    );
-}
