--- apps/src/lib/auth.ts
@@ -1,19 +1,0 @@
-// 1. src/lib/auth.ts - 토큰 관리 유틸리티
-export const getAuthToken = (): string | null => {
-    if (typeof window !== 'undefined') {
-        return localStorage.getItem('auth-token');
-    }
-    return null;
-};
-
-export const setAuthToken = (token: string): void => {
-    if (typeof window !== 'undefined') {
-        localStorage.setItem('auth-token', token);
-    }
-};
-
-export const removeAuthToken = (): void => {
-    if (typeof window !== 'undefined') {
-        localStorage.removeItem('auth-token');
-    }
-};
